LayoutTests/http/tests/websocket/tests/hybi/url-with-credential.html

<!DOCTYPE">
<html>
<head>
<script src="/js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Make sure handshake with URL with credentail info success.");

window.jsTestIsAsync = true;

var url = "ws://user:pass@127.0.0.1:8880/websocket/tests/hybi/echo-host";
var handshake_success = false;
var host;

function endTest()
{
    clearTimeout(timeoutID);
    shouldBeTrue("handshake_success");
    shouldBe("host", '"127.0.0.1:8880"');
    finishJSTest();
}

debug("url=" + url);
var ws = new WebSocket(url);
ws.onopen = function () {
    debug("WebSocket is open");
    handshake_success = true;
};
ws.onmessage = function (evt) {
    host = evt.data;
    debug("received:" + host);
    ws.close();
};
ws.onclose = function () {
    debug("WebSocket is closed");
    endTest();
};
var timeoutID = setTimeout("endTest()", 2000);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
