safari-600.7-branch/LayoutTests/http/tests/websocket/tests/hybi/url-with-credential-expected.txt

Make sure handshake with URL with credentail info success.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

url=ws://user:pass@127.0.0.1:8880/websocket/tests/hybi/echo-host
WebSocket is open
received:127.0.0.1:8880
WebSocket is closed
PASS handshake_success is true
PASS host is "127.0.0.1:8880"
PASS successfullyParsed is true

TEST COMPLETE
